dela upp stora klasser
