+++ app/src/main/groovy/se/alipsa/accounting/service/SieImportExportService.groovy
@@ -2,7 +2,6 @@ package se.alipsa.accounting.service
 
 import groovy.sql.GroovyRowResult
 import groovy.sql.Sql
-import groovy.transform.PackageScope
 
 import alipsa.sieparser.SIE
 import alipsa.sieparser.SieAccount
@@ -56,6 +55,7 @@ final class SieImportExportService {
   private final ReportIntegrityService reportIntegrityService
   private final AuditLogService auditLogService
   private final FiscalYearService fiscalYearService
+  private final SieImportJobRepository importJobRepository
   SieImportExportService() {
     this(
         DatabaseService.instance,
@@ -83,6 +83,7 @@ final class SieImportExportService {
     this.reportIntegrityService = reportIntegrityService
     this.auditLogService = auditLogService
     this.fiscalYearService = fiscalYearService
+    this.importJobRepository = new SieImportJobRepository(databaseService)
   }
 
   SieCompany peekSieCompany(Path filePath) {
@@ -113,7 +114,7 @@ final class SieImportExportService {
 
     databaseService.withSql { Sql sql ->
       FiscalYear existingYear = findFiscalYearByPeriod(sql, companyId, bookingYear.start, bookingYear.end)
-      ImportJob duplicate = findSuccessfulImportJobByChecksum(sql, companyId, checksum)
+      ImportJob duplicate = SieImportJobRepository.findSuccessfulByChecksum(sql, companyId, checksum)
       FiscalYearPurgeSummary purgeSummary = null
       List<String> blockingIssues = []
       if (existingYear != null && !(duplicate != null && !replaceExisting)) {
@@ -176,10 +177,10 @@ final class SieImportExportService {
     Path safePath = validateImportPath(filePath)
     byte[] content = Files.readAllBytes(safePath)
     String checksum = sha256(content)
-    long jobId = createImportJob(companyId, safePath.fileName.toString(), checksum)
+    long jobId = importJobRepository.create(companyId, safePath.fileName.toString(), checksum)
     // Duplicate-check is intentionally skipped for replacements: each replacement is a distinct destructive operation.
     if (!replaceExistingFiscalYear) {
-      ImportJob duplicate = markDuplicateIfNeeded(jobId, companyId, checksum)
+      ImportJob duplicate = importJobRepository.markDuplicateIfNeeded(jobId, companyId, checksum)
       if (duplicate != null) {
         return new SieImportResult(duplicate, null, true, 0, 0, 0, 0, [])
       }
@@ -208,7 +209,7 @@ final class SieImportExportService {
         String summary = replaceExistingFiscalYear
             ? buildReplacementSuccessSummary(counts, fiscalYear, parsed.warnings, purgeSummary)
             : buildSuccessSummary(counts, fiscalYear, parsed.warnings)
-        ImportJob job = completeImportJob(
+        ImportJob job = SieImportJobRepository.complete(
             sql,
             jobId,
             fiscalYear.id,
@@ -240,7 +241,7 @@ final class SieImportExportService {
       }
       result
     } catch (Exception exception) {
-      markFailed(jobId, resolvedFiscalYearId, exception)
+      importJobRepository.markFailed(jobId, resolvedFiscalYearId, exception)
       throw exception
     }
   }
@@ -283,26 +284,7 @@ final class SieImportExportService {
   }
   List<ImportJob> listImportJobs(long companyId, int limit = 20) {
     CompanyService.requireValidCompanyId(companyId)
-    int safeLimit = Math.max(1, limit)
-    databaseService.withSql { Sql sql ->
-      sql.rows('''
-          select ij.id,
-                 ij.file_name as fileName,
-                 ij.checksum_sha256 as checksumSha256,
-                 ij.fiscal_year_id as fiscalYearId,
-                 ij.status,
-                 ij.summary,
-                 ij.error_log as errorLog,
-                 ij.started_at as startedAt,
-                 ij.completed_at as completedAt
-            from import_job ij
-           where ij.company_id = ?
-           order by ij.started_at desc, ij.id desc
-           limit ?
-      ''', [companyId, safeLimit]).collect { GroovyRowResult row ->
-        mapImportJob(row)
-      }
-    }
+    importJobRepository.list(companyId, limit)
   }
   private static Path validateImportPath(Path filePath) {
     if (filePath == null) {
@@ -327,65 +309,6 @@ final class SieImportExportService {
     }
     normalized
   }
-  private long createImportJob(long companyId, String fileName, String checksum) {
-    databaseService.withTransaction { Sql sql ->
-      List<List<Object>> keys = sql.executeInsert('''
-          insert into import_job (
-              company_id,
-              file_name,
-              checksum_sha256,
-              status,
-              summary,
-              error_log,
-              started_at,
-              completed_at
-          ) values (?, ?, ?, 'STARTED', ?, null, current_timestamp, null)
-      ''', [companyId, truncate(fileName, 255), checksum, 'Import startad.'])
-      ((Number) keys.first().first()).longValue()
-    }
-  }
-  private ImportJob markDuplicateIfNeeded(long jobId, long companyId, String checksum) {
-    databaseService.withTransaction { Sql sql ->
-      GroovyRowResult existing = sql.firstRow('''
-          select id,
-                 file_name as fileName
-            from import_job
-           where checksum_sha256 = ?
-             and company_id = ?
-             and status = 'SUCCESS'
-             and id <> ?
-           order by completed_at desc, id desc
-           limit 1
-      ''', [checksum, companyId, jobId]) as GroovyRowResult
-      if (existing == null) {
-        return null
-      }
-      String summary = "Filen har redan importerats tidigare (jobb ${existing.get('id')})."
-      completeImportJob(sql, jobId, null, ImportJobStatus.DUPLICATE, summary, [])
-    }
-  }
-
-  private static ImportJob findSuccessfulImportJobByChecksum(Sql sql, long companyId, String checksum) {
-    GroovyRowResult row = sql.firstRow('''
-        select id,
-               file_name as fileName,
-               checksum_sha256 as checksumSha256,
-               fiscal_year_id as fiscalYearId,
-               status,
-               summary,
-               error_log as errorLog,
-               started_at as startedAt,
-               completed_at as completedAt
-          from import_job
-         where checksum_sha256 = ?
-           and company_id = ?
-           and status = 'SUCCESS'
-         order by completed_at desc, id desc
-         limit 1
-    ''', [checksum, companyId]) as GroovyRowResult
-    row == null ? null : mapImportJob(row)
-  }
-
   private static ParsedSie parseDocument(Path filePath) {
     SieDocumentReader reader = new SieDocumentReader()
     reader.throwErrors = false
@@ -1022,92 +945,6 @@ final class SieImportExportService {
     ByteArrayOutputStream output = new ByteArrayOutputStream()
     new SieDocumentWriter(document).write(output)
     output.toByteArray()
-  }
-
-  private static ImportJob completeImportJob(
-      Sql sql,
-      long jobId,
-      Long fiscalYearId,
-      ImportJobStatus status,
-      String summary,
-      List<String> warnings
-  ) {
-    String errorLog = warnings == null || warnings.isEmpty() ? null : truncate(warnings.join('\n'), 20000)
-    sql.executeUpdate('''
-        update import_job
-           set fiscal_year_id = ?,
-               status = ?,
-               summary = ?,
-               error_log = ?,
-               completed_at = current_timestamp
-         where id = ?
-    ''', [fiscalYearId, status.name(), truncate(summary, 1000), errorLog, jobId])
-    findImportJob(sql, jobId)
-  }
-
-  private void markFailed(long jobId, Long fiscalYearId, Exception exception) {
-    databaseService.withTransaction { Sql sql ->
-      sql.executeUpdate('''
-          update import_job
-             set fiscal_year_id = coalesce(?, fiscal_year_id),
-                 status = 'FAILED',
-                 summary = ?,
-                 error_log = ?,
-                 completed_at = current_timestamp
-           where id = ?
-      ''', [
-          fiscalYearId,
-          truncate(exception.message ?: 'Importen misslyckades.', 1000),
-          truncate(buildFailureLog(exception), 20000),
-          jobId
-      ])
-    }
-  }
-
-  private static String buildFailureLog(Throwable throwable) {
-    List<String> parts = []
-    Throwable current = throwable
-    while (current != null) {
-      if (current.message) {
-        parts << current.message
-      } else {
-        parts << current.class.simpleName
-      }
-      current = current.cause
-    }
-    parts.unique().join('\n')
-  }
-
-  private static ImportJob findImportJob(Sql sql, long jobId) {
-    GroovyRowResult row = sql.firstRow('''
-        select id,
-               file_name as fileName,
-               checksum_sha256 as checksumSha256,
-               fiscal_year_id as fiscalYearId,
-               status,
-               summary,
-               error_log as errorLog,
-               started_at as startedAt,
-               completed_at as completedAt
-          from import_job
-         where id = ?
-    ''', [jobId]) as GroovyRowResult
-    row == null ? null : mapImportJob(row)
-  }
-
-  @PackageScope
-  static ImportJob mapImportJob(GroovyRowResult row) {
-    new ImportJob(
-        Long.valueOf(row.get('id').toString()),
-        row.get('fileName') as String,
-        row.get('checksumSha256') as String,
-        row.get('fiscalYearId') == null ? null : Long.valueOf(row.get('fiscalYearId').toString()),
-        ImportJobStatus.valueOf(row.get('status') as String),
-        row.get('summary') as String,
-        SqlValueMapper.toClob(row.get('errorLog')),
-        SqlValueMapper.toLocalDateTime(row.get('startedAt')),
-        SqlValueMapper.toLocalDateTime(row.get('completedAt'))
-    )
   }
 
   private static FiscalYear findFiscalYear(Sql sql, long fiscalYearId) {
